create Toggle function to the WorkoutplansPage
