+++ src/Components/WorkoutplansPage/WorkouPlanMuscles.js
@@ -1,0 +1,11 @@
+import MusclesMap from './MusclesMap';
+
+export default function WorkoutPlanMuscles({ day }) {
+  return (
+    <>
+      {day.muscles.map(function (muscle, index) {
+        return <MusclesMap muscle={muscle} index={index} />;
+      })}
+    </>
+  );
+}
